requirements.txt
opencv-python
customtkinter
numpy
Pillow